Added in two new html files and fixed the video
The two new html files were for the Project page. To see the work I've done I've created a new page for each assignment for the user to see in the form of a slideshow. And for the video I've muted the raw clip and added a voice over of me explaining about myself